[#4] more tests
