+++ src/app/logic/generatingQuestions/gettingAnswer.js
